+++ econom_game/credits/models.py
@@ -7,7 +7,7 @@ from teams.models import Team
 class Credit(models.Model):
     team = models.ForeignKey(Team, related_name='credit_team', default=None)
     bank = models.ForeignKey(Bank, related_name='credit_bank', default=None)
-    debt_amount = models.PositiveIntegerField(default=0)
+    debt_amount = models.FloatField(default=0)
     term = models.PositiveIntegerField(default=1)
 
     def __str__(self):
